backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
email-validator
pydantic[email]
asyncpg
python-json-logger
loguru
